Add Cache implementation
`CacheRedditClient` is using the decorator pattern. It is implementing the `RedditClient` interface and accepts another `RedditClient` as a delegate. The `CacheReddit` implementation intercepts all calls and decides if it is going to handle it self or get the response from the delegate.

In `RedditServiceProvider`, we change the default singleton binding for `RedditService` to use the `CacheRedditClient` implementation. We then setup contextual binding to ensure `CacheRedditClient` will use the `GuzzleRedditClient` behind the scenes.

By default, Laravel will use the `file` cache driver and writes cache files to `storage/framework/cache`. Look at the contents of this directory before running anything new. Since we haven't done any cache work yet, it should be pretty empty. Visit one of the `/r/php`-style links and then look at the `storage/framework/cache` directory again. You should find a new file there. Look at thee contents of the file to get an idea what that looks like.

You should also notice that the first request to a new subreddit will be a bit slower than subsequent requests to that same subreddit. At least until the cache expires. :)

----

We *could* have implemented `Cache` in the following way (see below).

In doing so, we're making the Guzzle implemeentation too-aware of other concerns like cache. But using the decorator pattern, and using a `RedditClient` interface, we're able to focus all the work of actually getting the data in the `GuzzleRedditClient` but still able to fully leverage a cache system in the `CacheRedditClient`.

This point is driven home by the fact that neither `web.php` nor `GuzzleRedditClient` are modified in this change, yet the `/r/{subreddit}` functionality is fully leveraging cache.

Theere is nothing wrong with this implementation. It works. But it serves as a great way to show how/why you might want to consider other patterns when it comes to object design and dependency injection.

```php
class GuzzleRedditClient implements RedditClient
{
    private $client;
    private $cacheTimeInSeconds;

    /**
     * GuzzleRedditClient constructor.
     */
    public function __construct(Client $client, int $cacheTimeInSeconds = 300)
    {
        $this->client = $client;
        $this->cacheTimeInSeconds = $cacheTimeInSeconds;
    }

    public function getArticles(string $subreddit)
    {
        $key = sprintf('reddit.subreddit.getArticles.%s', $subreddit);

        return \Cache::remember($key, $this->cacheTimeInSeconds, function () use ($subreddit) {
            $url = sprintf('https://www.reddit.com/r/%s.json', $subreddit);

            $jsonResponse = $this->client->request('GET', $url);

            if ($jsonResponse->getStatusCode() !== 200) {
                abort($jsonResponse->getStatusCode());
            }

            $children = collect(json_decode(
                $jsonResponse->getBody()->getContents(),
                true
            )['data']['children']);

            return $children->pluck('data')->map(function($item) {
                return collect($item)->only([
                    'title',
                    'subreddit',
                    'url',
                    'selftext',
                    'created'
                ])->all();
            });
        });
    }
}
```